+++ docs-src/0.7/src/guides/cheat_sheet.md
