+++ packages/docs-router/src/lib.rs
@@ -157,6 +157,51 @@ pub fn CodeBlock(contents: String, name: Option<String>) -> Element {
     }
 }
 
+#[component]
+pub fn MermaidBlock(chart: &'static str) -> Element {
+    rsx! {
+        div {
+            document::Link { rel: "stylesheet", href: asset!("assets/mermaid_block.css") }
+            div {
+                class: "mermaid min-h-[60vh]",
+                style: "background-color: #fff; ",
+                "value": "{chart}",
+            }
+            script { r#type: "module",
+                r#"
+import mermaid from 'https://cdn.jsdelivr.net/npm/mermaid@11/+esm';
+import Panzoom from 'https://cdn.jsdelivr.net/npm/@panzoom/panzoom@4.6.0/+esm';
+
+mermaid.initialize({{
+    startOnLoad: false,
+}});
+
+const mermaidElements = document.querySelectorAll('.mermaid');
+let elements = [];
+mermaidElements.forEach((element, index) => {{
+    if (element.getAttribute('data-processed') === 'true') {{
+        return;
+    }}
+    element.textContent = element.value;
+    elements.push(element);
+}});
+
+mermaid.run().then(() => {{
+    elements.forEach((element, index) => {{
+        let svg = element.firstElementChild;
+        const panzoom = Panzoom(svg, {{
+            step: 1,
+            maxScale: 10,
+            minScale: 0.5,
+        }});
+        element.addEventListener('wheel', panzoom.zoomWithWheel);
+    }})
+}});"#
+            }
+        }
+    }
+}
+
 pub(crate) static Copy: Component<()> = |_| {
     rsx!(
         svg {
